+++ src/main/java/com/junho/recruitmentcollector/crawler/staticweb/StaticWebPageConnector.java
@@ -1,4 +1,4 @@
-package com.junho.recruitmentcollector.crawler;
+package com.junho.recruitmentcollector.crawler.staticweb;
 
 import com.junho.recruitmentcollector.exception.ConnectionFailedException;
 import org.jsoup.Jsoup;
